add table creation utility
